Add config_mode var into the task
Hi, 
I've added a variable to let to the user the opportunity to enter in config mode into the task. 
For some devices(e.g: Fortinet), several commands are only available in the config mode and it doesn't make any sense (from my point of vue) to use the task "send_config" just to send a simple command.